More tidy up of utils
