+++ lib/python/node_manager/utils/__init__.py
@@ -157,31 +157,6 @@ def release_branch_name(definition):
     )
 
 
-def node_type_name_from_components(definition, namespace=None, name=None, version=None):
-    """Generate a node type name based on its components.
-
-    Generate a new nodeTypeName for the given hou.HDADefinition updating the namespace,
-    name and version if provided.
-
-    Args:
-        definition(hou.HDADefinition): The HDA definition to generate the node type name
-            for.
-        namespace(:obj:`str`,optional): The new namespace to use for the definition.
-        name(:obj:`str`,optional): The new name to use for the definition.
-        version(:obj:`str`,optional): The new version to use for the definition.
-
-    Returns:
-        (str): The updated node type name.
-    """
-    current_name = definition.nodeTypeName()
-    new_namespace = nodetypeutils.node_type_namespace(current_name, new_namespace=namespace)
-    new_name = nodetypeutils.node_type_name(current_name, new_name=name)
-    new_version = nodetypeutils.node_type_version(current_name, new_version=version)
-    return "{namespace}::{name}::{version}".format(
-        namespace=new_namespace, name=new_name, version=new_version
-    )
-
-
 def editable_hda_path_from_components(definition, edit_dir, namespace=None, name=None):
     """Get the editable HDA path.
